requirements.txt
asgiref==3.2.7
Django==3.0.7
django-debug-toolbar==2.2
django-environ==0.4.5
pytz==2019.3
sqlparse==0.3.1
